%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  addPhenodata.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{addPhenodata}
\alias{addPhenodata}


\title{Gets bin annotation data for a particular bin size}

\usage{
addPhenodata(object, phenofile)
}

\description{
 Gets bin annotation data for a particular bin size.
}

\arguments{
  \item{object}{A \code{\link{QDNAseqReadCounts}} or \code{\link{QDNAseqCopyNumbers}}
  object.}
  \item{phenofile}{A file name with phenotypic data for samples in
  \code{object}.}
}

\value{
  Returns a named \code{\link[base]{list}} with element 'phenodata' ...
}

\author{Ilari Scheinin}

\seealso{
  ...
}


\keyword{IO}
