%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CGHcall.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{CGHcall}
\alias{CGHcall}


\alias{CGHcall,QDNAseqReadCounts-method}
\alias{CGHcall,cghSeg-method}

\title{Call aberrations from copy number data}

\usage{
CGHcall(object, ...)
}

\description{
 Call aberrations from copy number data.
}

\arguments{
  \item{object}{...}
  \item{...}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

