%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  compareToReference.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{compareToReference}
\alias{compareToReference}


\alias{compareToReference,QDNAseqReadCounts,numeric-method}

\title{Divide binned read counts with those of reference samples}

\usage{
compareToReference(object, references)
}

\description{
 Divide binned read counts with those of reference samples.
}

\arguments{
  \item{object}{...}
  \item{references}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

