%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  compareToReferenceReadCounts.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{compareToReferenceReadCounts}
\alias{compareToReferenceReadCounts}


\title{Divide binned read counts with those of reference samples}

\usage{
compareToReferenceReadCounts(obj, references)
}

\description{
 Divide binned read counts with those of reference samples.
}

\arguments{
  \item{obj}{...}
  \item{references}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

