%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  getBins.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{createBins}
\alias{createBins}


\title{Builds bin annotation data for a particular bin size}

\usage{
createBins(binsize, genome="hg19")
}

\description{
 Builds bin annotation data for a particular bin size.
}

\arguments{
  \item{binsize}{A \code{\link[base]{numeric}} scalar specifying ...}
  \item{genome}{A \code{\link[base]{character}} string ...}
}

\value{
  Returns ...
}

\author{Ilari Scheinin}

\seealso{
  \code{\link{getBins}}().
}

