%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ExpandCGHcall.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{ExpandCGHcall}
\alias{ExpandCGHcall}


\alias{ExpandCGHcall,list,QDNAseqReadCounts-method}
\alias{ExpandCGHcall,list,cghSeg-method}

\title{Expands result from CGHcall to a QDNAseqReadCounts or CGHcall object}

\usage{
ExpandCGHcall(listcall, object, ...)
}

\description{
 Expands result from CGHcall to a QDNAseqReadCounts or CGHcall object.
}

\arguments{
  \item{listcall}{...}
  \item{object}{...}
  \item{...}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

