%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  filterBins.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{filterBins}
\alias{filterBins}


\title{Adjusts the filter on which bins are used}

\usage{
filterBins(obj, blacklist=0, mappability=50, tgr=2, bases=100, allosomes=TRUE)
}

\description{
 Adjusts the filter on which bins are used.
}

\arguments{
  \item{obj}{A qdnaseq object ...}
  \item{blacklist}{...}
  \item{mappability}{...}
  \item{tgr}{...}
  \item{bases}{...}
  \item{allosomes}{...}
}

\value{
  Returns a named \code{\link[base]{list}} with element 'phenodata' ...
}

\author{Ilari Scheinin}

\seealso{
  ...
}


\keyword{IO}
