%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  getBins.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBins}
\alias{getBins}


\title{Gets bin annotation data for a particular bin size}

\usage{
getBins(binsize, genome="hg19", cache=TRUE, force=FALSE)
}

\description{
 Gets bin annotation data for a particular bin size.
}

\arguments{
  \item{binsize}{A \code{\link[base]{numeric}} scalar specifying ...}
  \item{genome}{A \code{\link[base]{character}} string ...}
  \item{cache}{A \code{\link[base]{logical}} ...}
  \item{force}{A \code{\link[base]{logical}} ...}
}

\value{
  Returns ...
}

\author{Ilari Scheinin}

\seealso{
  \code{\link{createBins}}.
}


\keyword{IO}
