%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  plot-methods.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isobarPlot}
\alias{isobarPlot}


\alias{isobarPlot,QDNAseqReadCounts,missing-method}

\title{Plot median read counts as a function of GC content and mappability}

\usage{
isobarPlot(x, y, ...)
}

\description{
 Plot median read counts as a function of GC content and mappability.
}

\arguments{
  \item{x}{...}
  \item{y}{...}
  \item{...}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

