%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  998.LGG150.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{LGG150}
\alias{LGG150}
\docType{data}


 \title{LGG150 chromosome 9}

 \description{
   LGG150 chromosome 9
 }

 \author{Ilari Scheinin}
\keyword{datasets}
