%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  makeCgh.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{makeCgh}
\alias{makeCgh}


\alias{makeCgh,QDNAseqReadCounts-method}

\title{Constructs a 'cghRaw', 'cghSeg', or 'cghCal' object}

\usage{
makeCgh(object)
}

\description{
 Constructs a 'cghRaw', 'cghSeg', or 'cghCal' object.
}

\arguments{
  \item{object}{...}
}

\value{
  Returns a \code{\link[CGHbase]{cghRaw}} object.
}

\author{Ilari Scheinin}

