%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  makeCghRaw.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{makeCghRaw}
\alias{makeCghRaw}


\title{Constructs a 'cghRaw' object}

\usage{
makeCghRaw(obj)
}

\description{
 Constructs a 'cghRaw' object.
}

\arguments{
  \item{obj}{...}
}

\value{
  Returns a \code{\link[CGHbase]{cghRaw}} object.
}

\author{Ilari Scheinin}

\seealso{
  Internally, \code{\link[CGHbase]{make_cghRaw}} is used.
}

