%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  plot-methods.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{noisePlot}
\alias{noisePlot}


\alias{noisePlot,QDNAseqReadCounts,missing-method}

\title{Plot noise as a function of sequence depth}

\usage{
noisePlot(x, y, ...)
}

\description{
 Plot noise as a function of sequence depth.
}

\arguments{
  \item{x}{...}
  \item{y}{...}
  \item{...}{...}
}

\author{Ilari Scheinin}

