%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  normalizeReadCounts.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{normalizeReadCounts}
\alias{normalizeReadCounts}


\title{Normalizes binned read counts}

\usage{
normalizeReadCounts(obj, method="median", logTransform=TRUE, smoothOutliers=TRUE)
}

\description{
 Normalizes binned read counts.
}

\arguments{
  \item{obj}{...}
  \item{method}{A \code{\link[base]{character}} string specifying ...}
  \item{logTransform}{If \code{\link[base:logical]{TRUE}}, ..., otherwise, ...}
  \item{smoothOutliers}{If \code{\link[base:logical]{TRUE}}, ..., otherwise, ...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

