%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  postsegnormalize.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{postsegnormalize}
\alias{postsegnormalize}


\alias{postsegnormalize,QDNAseqReadCounts-method}
\alias{postsegnormalize,cghSeg-method}

\title{Post-segmentation normalization}

\usage{
postsegnormalize(object, ...)
}

\description{
 Post-segmentation normalization.
}

\arguments{
  \item{object}{...}
  \item{...}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

