\name{QDNAseq-defunct}
\alias{downloadBinAnnotations}
\title{Defunct functions in package \sQuote{QDNAseq}}

\description{
    These functions are defunct and no longer available.
}

\details{

    The following functions are defunct; use the replacement indicated below:
    \itemize{
        \item{\code{downloadBinAnnotations}: \code{\link{getBinAnnotations}}}
    }
}
