\name{QDNAseq-deprecated}
\alias{QDNAseq-deprecated}
\title{Deprecated functions in package \sQuote{QDNAseq}}

\description{
    These functions are provided for compatibility with older versions
    of \sQuote{QDNAseq} only, and will be defunct at the next release.
}

\details{

    The following functions are deprecated and will be made defunct; use
    the replacement indicated below:
    \itemize{

        \item{downloadBinAnnotations: \code{\link{getBinAnnotations}}}

    }
}
