%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{qdnaseq-package}
\alias{qdnaseq-package}
\alias{qdnaseq}
\docType{package}

\title{Package qdnaseq}


\description{
  More about what it does (maybe more than one line)
}

\section{Requirements}{
  TO DO.
}

\section{How to cite this package}{
  Whenever using this package, please cite

  \emph{No citation information available.}

}

\author{Ilari Scheinin}

\section{License}{
  The releases of this package is licensed under GPL.
}
\keyword{package}
