%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QDNAseqCopyNumbers-accessors.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{QDNAseqCopyNumbers}
\docType{class}
\alias{QDNAseqCopyNumbers}
\alias{QDNAseqCopyNumbers-class}


\title{Container for QDNAseq read count data}

\description{
  Container for QDNAseq read count data
}

\section{Assay data elements}{
  An object of this class contains the following elements:
  \describe{
    \item{\code{copynumber}}{(\code{\link[base]{numeric}}) Corrected
      "count" signals in \eqn{[0,+\infty)}}.
  }
}

\section{Missing values}{
  The bin data (assay data) may contain missing values.
}

\author{Ilari Scheinin}
\keyword{classes}
