%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QDNAseqSignals-accessors.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{QDNAseqSignals}
\docType{class}
\alias{QDNAseqSignals}
\alias{QDNAseqSignals-class}


\alias{chromosomes,QDNAseqSignals-method}
\alias{bpstart,QDNAseqSignals-method}
\alias{bpend,QDNAseqSignals-method}

\title{A parent class for containers of QDNAseq data}

\description{
    A parent class for containers of QDNAseq data
}

\author{Ilari Scheinin}
\keyword{classes}
