%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  binReadCounts.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{.binReadCountsPerSample}
\alias{.binReadCountsPerSample}


\title{Calculate binned read counts from a BAM file}

\usage{
.binReadCountsPerSample(bins, bamfile, path, cache=TRUE, samtools="samtools", f="",
  F="0x0404", q=37, maxChunk=100000000L)
}

\description{
 Calculate binned read counts from a BAM file.
}

\arguments{
  \item{bins}{...}
  \item{bamfile}{...}
  \item{path}{...}
  \item{cache}{...}
  \item{samtools}{...}
  \item{f}{...}
  \item{F}{...}
  \item{q}{...}
  \item{maxChunk}{...}
}

\value{
  Returns ...
}

\author{Ilari Scheinin}

\seealso{
  To retrieve bin read counts for a set of BAM files,
  see \code{\link{binReadCounts}}().
}


\keyword{IO}
