%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  binReadCounts.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{.binReadCountsPerSample}
\alias{.binReadCountsPerSample}


\title{Calculate binned read counts from a BAM file}

\usage{
.binReadCountsPerSample(bins, bamfile, path, cache, samtools, f, F, q, maxChunk)
}

\description{
 Calculate binned read counts from a BAM file.
}

\arguments{
  \item{bins}{...}
  \item{bamfile}{...}
  \item{path}{...}
  \item{cache}{...}
  \item{samtools}{...}
  \item{f}{...}
  \item{F}{...}
  \item{q}{...}
  \item{maxChunk}{...}
}

\value{
  Returns ...
}

\author{Ilari Scheinin}

\seealso{
  To retrieve bin read counts for a set of BAM files,
  see \code{\link{binReadCounts}}().
}


\keyword{IO}
