%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  addPhenodata.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{addPhenodata}
\alias{addPhenodata}


\title{Gets bin annotation data for a particular bin size}

\usage{
addPhenodata(obj, phenofile)
}

\description{
 Gets bin annotation data for a particular bin size.
}

\arguments{
  \item{obj}{A named \code{\link[base]{list}} containing ...}
  \item{phenofile}{A \code{\link[base:connections]{connection}} of a filename specifying ...}
}

\value{
  Returns a named \code{\link[base]{list}} with element 'phenodata' ...
}

\author{Ilari Scheinin}

\seealso{
  ...
}


\keyword{IO}
