%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  applyFilters.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{applyFilters}
\alias{applyFilters}


\alias{applyFilters,QDNAseqReadCounts-method}

\title{Adjusts the filter on which bins are used}

\usage{
applyFilters(object, mappability=50, blacklist=0, residual=1, bases=100,
  filterAllosomes=TRUE, force=FALSE)
}

\description{
 Adjusts the filter on which bins are used.
}

\arguments{
  \item{object}{A QDNAseqReadCounts object ...}
  \item{mappability}{...}
  \item{blacklist}{...}
  \item{residual}{...}
  \item{bases}{...}
  \item{filterAllosomes}{...}
  \item{force}{...}
}

\value{
  Returns a named \code{\link[base]{list}} with element 'phenodata' ...
}

\author{Ilari Scheinin}

\seealso{
  ...
}


\keyword{IO}
