%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  plot-methods.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{frequencyPlot}
\alias{frequencyPlot}


\alias{frequencyPlot,QDNAseqReadCounts,missing-method}

\title{Plot copy number aberration frequencies}

\usage{
frequencyPlot(x, y, ...)
}

\description{
 Plot copy number aberration frequencies.
}

\arguments{
  \item{x}{...}
  \item{y}{...}
  \item{...}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

