%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  highlightFilters.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{highlightFilters}
\alias{highlightFilters}


\alias{highlightFilters,QDNAseqReadCounts-method}

\title{Highlights data points in a plotted profile to evaluate filtering}

\usage{
highlightFilters(object, col="red", residual=NA, blacklist=NA, mappability=NA, bases=NA,
  type="union", ...)
}

\description{
 Highlights data points in a plotted profile to evaluate filtering.
}

\arguments{
  \item{object}{A QDNAseqReadCounts object ...}
  \item{col}{...}
  \item{residual}{...}
  \item{blacklist}{...}
  \item{mappability}{...}
  \item{bases}{...}
  \item{type}{...}
  \item{...}{Further arguments to points.}
}

\author{Ilari Scheinin}

\seealso{
  ...
}


\keyword{IO}
