%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  highlightFilters.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{highlightFilters}
\alias{highlightFilters}


\title{Highlights data points in a plotted profile to evaluate filtering}

\usage{
highlightFilters(object, col="red", mappability=50, blacklist=0, residual=1, bases=100,
  ...)
}

\description{
 Highlights data points in a plotted profile to evaluate filtering.
}

\arguments{
  \item{object}{A QDNAseqReadCounts object ...}
  \item{col}{...}
  \item{mappability}{...}
  \item{blacklist}{...}
  \item{residual}{...}
  \item{bases}{...}
  \item{...}{Further arguments to points.}
}

\author{Ilari Scheinin}

\seealso{
  ...
}


\keyword{IO}
