%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  998.LGG150.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{LGG150}
\alias{LGG150}
\docType{data}


 \title{LGG150 chromosomes 7-10}

 \description{
   An example data set of read counts from chromosomes 7-10 of sample
   LGG150, contained within a \code{\link{QDNAseqReadCounts}} object
 }

 \author{Ilari Scheinin}
\keyword{datasets}
