%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  poolRuns.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{poolRuns}
\alias{poolRuns}


\alias{poolRuns,QDNAseqSignals,character-method}

\title{Pools binned read counts across samples}

\usage{
poolRuns(object, samples, force=FALSE)
}

\description{
 Pools binned read counts across samples.
}

\arguments{
  \item{object}{...}
  \item{samples}{...}
  \item{force}{...}
}

\value{
  Returns a named \code{\link[base]{list}} containing elements ...
}

\author{Ilari Scheinin}

\seealso{
  Internally, ...
}

