%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{QDNAseq-package}
\alias{QDNAseq-package}
\alias{QDNAseq}
\docType{package}

\title{Package QDNAseq}


\description{
    Quantitative DNA sequencing for chromosomal aberrations.
}

\section{Details}{
    A package to detect chromosomal aberrations from whole-genome sequencing
    data. \code{\link{QDNAseqReadCounts}} and \code{\link{QDNAseqCopyNumbers}} classes are
    used as the main data structures.
}

\section{How to cite this package}{
    Whenever using this package, please cite:
    Scheinin I, Sie D, Bengtsson H, van de Wiel MA, Olshen AB, van Thuijl HF, van Essen HF, Eijk PP, Rustenburg F, Meijer GA, Reijneveld JC, Wesseling P, Pinkel D, Albertson DG and Ylstra B
(2014). "DNA copy number analysis of fresh and formalin-fixed specimens by shallow whole-genome sequencing with identification and exclusion of problematic regions in the genome assembly."
_Genome Research_. doi:10.1101/gr.175141.114.
}

\author{Ilari Scheinin}

\section{License}{
    This package is licensed under GPL.
}
\keyword{package}
