%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{QDNAseq-package}
\alias{QDNAseq-package}
\alias{QDNAseq}
\docType{package}

\title{Package QDNAseq}


\description{
  Quantitative DNA sequencing for chromosomal aberrations.
}

\section{Details}{
  A package to detect chromosomal aberrations from whole-genome sequencing
  data. \code{\link{QDNAseqReadCounts}} and \code{\link{QDNAseqCopyNumbers}} classes are
  used as the main data structures.
}

\section{How to cite this package}{
  Whenever using this package, please cite
  \preformatted{
Ilari Scheinin (2014). QDNAseq: Quantitative DNA sequencing for chromosomal
aberrations. R package version 0.7.6.
}
\emph{}

}

\author{Ilari Scheinin}

\section{License}{
  The releases of this package is licensed under GPL.
}
\keyword{package}
