%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AllClasses.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{QDNAseqReadCounts}
\docType{class}
\alias{QDNAseqReadCounts}
\alias{QDNAseqReadCounts-class}


 \alias{chromosomes,QDNAseqReadCounts-method}
 \alias{bpstart,QDNAseqReadCounts-method}
 \alias{bpend,QDNAseqReadCounts-method}

 \title{Container for qdnaseq data}

 \description{
   Container for qdnaseq data
 }

 \author{Ilari Scheinin}
\keyword{classes}
